Add option to install to updates subdir
Sometimes you'd rather have kmodtool write the kmod files to the updates
folder of the kernel modules directory.

An example for this case would be any module that is installed by the
kernel-modules-extra package (which installs all it's modules to the
extra subfolder), because any kmod package that would also install
itself to the extra directory would not overload the existing modules
in that same directory.

Signed-off-by: Felix Kaechele <felix@kaechele.ca>